"Rest configuration"

+++ src/main/java/br/com/ecommerce/productAPI/controller/GenericController.java
@@ -9,6 +9,7 @@ import br.com.ecommerce.productAPI.service.GenericService;
 import java.io.Serializable;
 import java.util.List;
 import javax.validation.Valid;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -24,37 +25,50 @@ import org.springframework.web.bind.annotation.RequestBody;
  * @param <ID>
  */
 public abstract class GenericController<T, ID extends Serializable> {
-    
+
     protected abstract GenericService<T, ID> getService();
-    
+
     @GetMapping(value = "/{id}")
-    public ResponseEntity<T> findOne(@PathVariable ID id) {
+    public ResponseEntity findOne(@PathVariable ID id) {
         T entity = getService().findOne(id);
-        return ResponseEntity.ok(entity);
+        if (entity != null) {
+            return new ResponseEntity<>(entity, HttpStatus.CREATED);
+        }
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Não encontrado");
     }
-    
+
     @GetMapping
-    public ResponseEntity<List<T>> findAll(){
+    public ResponseEntity<List<T>> findAll() {
         List<T> list = getService().findAll();
         return ResponseEntity.ok(list);
     }
-    
+
     @PostMapping
-    public ResponseEntity<T> save(@Valid @RequestBody T entity){
+    public ResponseEntity save(@Valid @RequestBody T entity) {
         T created = getService().save(entity);
-        return ResponseEntity.ok(created);
+        if (created != null) {
+            return new ResponseEntity<>(created, HttpStatus.CREATED);
+        }
+        return ResponseEntity
+                .status(HttpStatus.CONFLICT)
+                .body("Ocorreu um erro na requisição");
     }
-    
+
     @PutMapping
-    public ResponseEntity<T> update(@Valid @RequestBody T entity){
-        T created = getService().save(entity);
-        return ResponseEntity.ok(created);
+    public ResponseEntity<T> update(@Valid @RequestBody T entity) {
+        return save(entity);
     }
-    
+
     @DeleteMapping(value = "/{id}")
-    public ResponseEntity<String> delete(@PathVariable ID id){
-        getService().delete(id);
-        return ResponseEntity.ok("It was Deleted");
+    public ResponseEntity delete(@PathVariable ID id) {
+        T found = getService().findOne(id);
+        if(found != null){
+            getService().delete(id);
+            return ResponseEntity.ok().build();
+        }
+        return ResponseEntity
+                .status(HttpStatus.NOT_FOUND)
+                .body("Não foi deletar");
     }
-    
+
 }
